chore: update .gitignore and package.json
Update .gitignore to remove the exclusion of .npmrc file and remove the unnecessary line break at the end of the file.

Add .npmrc file to the project, which sets the registry for the @nomble package to be the GitHub package registry.

Update package.json to change the registry URL in the publishConfig to the GitHub package registry URL.

These changes are made to configure the project to use the GitHub package registry for the @nomble package.